+++ project1/application.py
@@ -5,17 +5,12 @@ from models import *
 from datetime import datetime
 from sqlalchemy import and_
 from book_details import get_book_details
+from init_app import initialize_production_app
 
-app = Flask(__name__)
-app.secret_key = 'anyrandomstring'
-# Check for environment variable
-if not os.getenv("DATABASE_URL"):
-    raise RuntimeError("DATABASE_URL is not set")
-
-# app.config["SQLALCHEMY_DATABASE_URI"] = "postgres://tgsemzadyezlgh:858baa7b0d8b5ce3cb37d6481e187d508a119f33d2f26a2addac896f186eb633@ec2-18-233-137-77.compute-1.amazonaws.com:5432/d389gkjdhv6oa2"
-app.config["SQLALCHEMY_DATABASE_URI"] = os.getenv("DATABASE_URL")
-app.config["SQLALCHEMY_TRACK_MODIFICATIONS"] = False
-db.init_app(app)
+try:
+    app = initialize_production_app()
+except Exception as e:
+    print("Error: Check for Database url with respect to configurations...")
 
 @app.route("/")
 def index():
